feat: lazily require modules on ad-hoc basis
